requirements.txt
dj-rest-auth==2.2.4
Django==4.0.4
django-allauth==0.50.0
djangorestframework==3.13.1
